--- .vscode/settings.json
@@ -1,5 +1,0 @@
-{
-	"python.analysis.stubPath": "c:\\Users\\mariu\\AppData\\Roaming\\Code\\User\\globalStorage\\vexrobotics.vexcode\\sdk\\python\\IQ2\\IQ2_1_0_0_21\\vexiq2\\stubs",
-	"python.analysis.diagnosticMode": "workspace",
-	"python.analysis.typeCheckingMode": "basic"
-}
