--- .vscode/vex_project_settings.json
@@ -1,18 +1,0 @@
-{
-	"extension": {
-		"version": "0.7.2025041600",
-		"json": 2
-	},
-	"project": {
-		"name": "Wall-e",
-		"description": "",
-		"creationDate": "10/2/2025, 8:49:41 PM",
-		"platform": "IQ2",
-		"language": "python",
-		"slot": 1,
-		"sdkVersion": "IQ2_1_0_0_21",
-		"python": {
-			"main": "automation_all_in_one.py"
-		}
-	}
-}
