+++ automation_all_in_one.py
@@ -5,7 +5,7 @@ import time
 brain = Brain()
 
 # PARAMETERS
-DISTANCE_THRESHOLD = 300   # mm
+DISTANCE_THRESHOLD = 400   # mm
 DRIVE_SPEED = 40 # maybe more?
 GRAB_SPEED = 60 # maybe less?
 SLOT_DISTANCE_DEGREES = 360 # needs to be measured, how many rolls is each "cube" on the field
@@ -13,20 +13,17 @@ RETURN_SPEED = 40
 MAX_RETRIES = 2
 
 # PORTS
-motor_gripper = Motor(Ports.PORT6)
+motor_gripper = Motor(Ports.PORT5)
 motor_ejecter = Motor(Ports.PORT7)
 motor_trolley = Motor(Ports.PORT8)
-motor_thumb = Motor(Ports.PORT9)
 distance_sensor = Distance(Ports.PORT4)
 bumper_sensor = Bumper(Ports.PORT10)
-optical_sensor = Optical(Ports.PORT12)
-touch_sensor = Touchled(Ports.PORT11)
 
-# ALL in degrees, test dummy vals, distance from 0
-SLOT_BLUE_DEG = 950
-SLOT_RED_DEG = 300
-SLOT_GREEN_DEG = 100
-DIST_TO_HARVESTING_SITE = 1300
+# ALL in degrees, test dummy vals
+SLOT_BLUE_DEG = 400
+SLOT_RED_DEG = 1000
+SLOT_BLUE_DEG = 700
+DIST_TO_HARVESTING_SITE = 1500
 
 # ----------------- SETUP (already given) -----------------
 def setup():
@@ -36,11 +33,26 @@ def setup():
     # PORTS
 
 
+    try:
+        touch_sensor = Touchled(SENSOR_TOUCH_PORT)
+        print("TouchLED OK")
+    except:
+        print("TouchLED Error")
+
+    try:
+        optical_sensor = Optical(SENSOR_OPTICAL_PORT)
+        print("Color OK")
+    except:
+        optical_sensor = None
+        print("Color Error")
+
+
+
+
 
     motor_gripper.position(DEGREES)
     motor_trolley.position(DEGREES)
     motor_ejecter.position(DEGREES)
-    motor_thumb.position(DEGREES)
 
 
 
@@ -151,7 +163,6 @@ def set_motor_torque(motor, torque_percent):
                 motor.set_torque(abs(torque_percent), PERCENT)
                 motor.spin(FORWARD)
             elif torque_percent < 0:
-                print("sett backwawrd torque")
                 motor.set_torque(abs(torque_percent), PERCENT)
                 motor.spin(REVERSE)
             else:
@@ -195,22 +206,14 @@ def grab_cube():
     start_time = time.time()
     delta_time = 0
     success = True
-    while get_distance() > 14:
-        print("retrieving cube")
-        print(get_distance())
-        spin_motor(motor_gripper, -GRAB_SPEED)
+    while get_distance() > 1:
+        spin_motor(motor_gripper, GRAB_SPEED)
         delta_time = time.time() - start_time
-        time.sleep(0.07)
-        if delta_time > 6:  # timeout after 6 seconds
+        if delta_time > 6000:  # timeout after 4 seconds
             print("Timeout reached while grabbing cube.")
-            if get_distance() < 25: # tbd deal with error handling
-                success = True
-            else:
-                success = False
+            success = False
             break
     spin_motor(motor_gripper, 0)
-    print("FINISHED GRABBING, SUCCESS:")
-    print(success)
     return success
     # think about error handling ######################
 
@@ -218,38 +221,23 @@ def grab_cube():
 
 def release_gripper():
     print("moving gripper all the way in front")
-    start_time = time.time()
+    start_time = 0
     delta_time = 0
-    while delta_time < 2.5:
-        spin_motor(motor_gripper, GRAB_SPEED)
+    while delta_time < 4000:
+        spin_motor(motor_gripper, -GRAB_SPEED)
         delta_time = time.time() - start_time
     # move all the way to the maximum
     spin_motor(motor_gripper, 0)
-    print("Gripper moved to end pos")
 
 
 
 def release_cube():
-   # print("Releasing cube...")
-   # start_time = time.time()
-   # delta_t = 0
-   # print(time.time())
-    # move gripper back to allow ejecter to eject cube
-   # while delta_t < 2.5:
-   #     spin_motor(motor_gripper, GRAB_SPEED)
-   #     delta_t = time.time() - start_time
-   # print(time.time())
-    
-   # stop_motor(motor_gripper)
-   # print("Gripper moved to end pos")
-    
-    open_thumb()
-    time.sleep(0.3)
-    print("ejecting cube")
-    spin_motor_to_position(motor_ejecter,-500,-100)
+    print("Releasing cube...")
+    spin_motor(motor_1, -GRAB_SPEED)
+    spin_motor(motor_2, -GRAB_SPEED)
     wait(500, MSEC)
-    print("ejected cube")
-    spin_motor_to_position(motor_ejecter,0,100)
+    stop_motor(motor_1)
+    stop_motor(motor_2)
 
 def grab_with_retry():
     for attempt in range(1, MAX_RETRIES + 1):
@@ -265,28 +253,6 @@ def grab_with_retry():
     print("All grab attempts failed. Skipping cube.")
     return False
 
-def open_thumb():
-    print("Opening thumb...")
-    set_motor_torque(motor_thumb, 0)
-    start_time = time.time()
-    delta_time = 0
-    while delta_time < 0.24:
-        spin_motor(motor_thumb, 60)
-        delta_time = time.time() - start_time
-    spin_motor(motor_thumb, 0)
-
-
-def close_thumb():
-
-    print("Closing thumb...")
-    start_time = time.time()
-    delta_time = 0
-    while delta_time < 1:
-        spin_motor(motor_thumb, -100)
-        delta_time = time.time() - start_time
-    set_motor_torque(motor_thumb, -100)
-    
-
 
 
 def return_to_start():
@@ -298,20 +264,15 @@ def return_to_start():
     stop_motor(motor_4)
 
 # --------- Sorting ---------
-def place_cube(color, absolute_pos):
-    print("Placing cube:")
-    print(color)
+def place_cube(color, rel_pos):
+    print("Placing cube:", color)
     if color == "green":
-        spin_motor_to_position(motor_trolley, -SLOT_GREEN_DEG, 100)
+        spin_motor_to_position(motor_trolley, rel_pos + SLOT_GREEN_DEG, 80)
     elif color == "red":
-        spin_motor_to_position(motor_trolley, -SLOT_RED_DEG, 100)
+        spin_motor_to_position(motor_trolley, rel_pos + SLOT_RED_DEG, 80)
     elif color == "blue":
-        spin_motor_to_position(motor_trolley, -SLOT_BLUE_DEG, 100)
+        spin_motor_to_position(motor_trolley, rel_pos + SLOT_BLUE_DEG, 80)
         print("Dropping cube into trash chute")
-    elif color == "none":
-        print("Unknown color, placing in trash chute")
-        spin_motor_to_position(motor_trolley, -SLOT_BLUE_DEG, 100)
-    
     release_cube()
     wait(300, MSEC)
 
@@ -338,91 +299,55 @@ def autonomous_run():
     print("System ready")
 
     # start from centered at base, arm fully extended
-    absolute_pos = DIST_TO_HARVESTING_SITE
-    start_time_run = time.time()
-    passed_time = 0
+    rel_pos = DIST_TO_HARVESTING_SITE
 
-    while passed_time < 600: # tbd more discrete timing
-        spin_motor_to_position(motor_trolley, -absolute_pos, 100) # move to mining area
-        print("spinned_motor to first pos")
-        mov_step = 75
-        # look for cube
-        print("checking for cube...")
+    spin_motor_to_position(motor_trolley, -rel_pos, 100) # move to mining area
+    print("spinned_motor to first pos")
+    mov_step = 10
+    # look for cube
+    print("checking for cube...")
+    # debug
+    while not bumper_pressed():
+        time.sleep(0.05)
 
-        time_out_time = 22
-        start_time_outtimer = time.time()
-        d_t = 0
-        success = False
-        while d_t < time_out_time: # add break point
-            spin_motor_to_position(motor_trolley, -absolute_pos - mov_step, 10)
+    while True:
+        spin_motor_to_position(motor_trolley, -rel_pos - mov_step, 10)
 
-            dist = get_distance()
-            print("Distance: ")
-            print(dist)
+        dist = get_distance()
+        print("Distance: ")
 
-            absolute_pos += mov_step
-            if 0 < dist < DISTANCE_THRESHOLD:
-                success = True
-                break
-            d_t = time.time() - start_time_outtimer
-        if passed_time >= 590:
-            spin_motor(motor_trolley,0)
+        print(dist)
+        rel_pos += mov_step
+        if 0 < dist < DISTANCE_THRESHOLD:
             break
-        if not success:
-            while not bumper_pressed():
-                spin_motor(motor_trolley,100)
-                motor_trolley.set_position(0, DEGREES)
-            continue
-        # move distance so cube is centered
-        spin_motor_to_position(motor_trolley, -absolute_pos - 150, 100) # old -180
 
-        if passed_time >= 590: # stop timing
-            spin_motor(motor_trolley,0)
+    # debug
+    while not bumper_pressed():
+        time.sleep(0.05)
 
-        # cube detected, retrieve it
-        if grab_cube(): # success
-            # get color
-            print("cube aligned with cart")
-            set_color_sensor_led(True)
-            time.sleep(0.2)
-            color = get_color()
-            print(color)
-            ctr = 0
-            
-            if passed_time >= 590: # stop timing
-                spin_motor(motor_trolley,0)
-            while ctr < 5:
-                color_test = get_color()
-                ctr += 1
-                if color_test != color:
-                    ctr = 0
-                print(color_test)
-            color = color_test
-            set_color_sensor_led(False)
-            set_led(color)
+    # cube detected, retrieve it
+    if grab_cube(): # success
+        # get color
+        color = get_color()
+        #print(f"Cube detected, color: {color}, Distance: {dist}mm")
+        place_cube(color, rel_pos)
 
-            #release the gripper immediately after grabbing the cube
+    else: # failed
+        print("Failed to grab cube, aborting mission.")
+    
 
-            release_gripper()
-            close_thumb()
-                
-            if passed_time >= 590: # stop timing
-                spin_motor(motor_trolley,0)
-            
-            #print(f"Cube detected, color: {color}, Distance: {dist}mm")
-
-            place_cube(color, absolute_pos)
-            set_led("off")
-
-        else: # failed
-            release_gripper()
-            print("Failed to grab cube, aborting mission.")
-        
-        while not bumper_pressed():
-            spin_motor(motor_trolley,100)
-        motor_trolley.set_position(0, DEGREES)
-        passed_time = time.time() - start_time_run
-
+        #     color = get_color()
+        #     print(f"Cube detected, color: {color}, Distance: {dist}cm")
+        #     if grab_with_retry():
+        #         return_to_start()
+        #         place_cube(color)
+        #         return_to_start()
+        #     else:
+        #         print("Skipping cube after failed grabs.")
+        # else:
+        #     spin_motor(motor_3, DRIVE_SPEED)
+        #     spin_motor(motor_4, DRIVE_SPEED)
+        #     wait(100, MSEC)
 
 
 autonomous_run()
